[Sailfish UI] use busy state on reporterSwitch when reporter is starting
When crash reporter launches for the first time after reboot, there is
30 seconds delay period before the service appears on D-Bus and starts
doing anything, in order not to slow down system startup with upload of
accumulated reports.

When this first startup was done via settings applet, there used to be
a long interim between tapping the switch and the bullet lighting up,
which might have been confusing for the user. Now the bullet blinks
during the daemon's activating period, giving an immediate feedback.