[endurance] don't depend on btime in /proc/stat to recogize a reboot
The value is can fluctuate few seconds over longer periods of time,
particulary when device has automatic update of date & time enabled.
This confused endurance-collect's notion of sessions, causing it to
prematurely end a session even though no reboot was done.

We now create a file in /tmp, whose absence after it gets wiped with
reboot marks the start of a new session.