[crashreporter] Avoid writing temporary files to fs when packing endurance snapshot. JB#57243
Instead of unpacking every lzofile to disk and creating a tar from this
use the knowledge of tar archives as a base for the packing operation.

It is not possible to extend tar from stdin, as tar requires information
about the filesystem to create it's structures. But we can do the tar
incrementally by generating the tar structure ourself.

Tar has 512 bytes header, data block, and padding to get even 512 blocks
of data.

1. Create empty hollow file to represent the file stored in the tar package
2. Create only header from that file - send it to stdout
3. Unpack the lzo file to stdout (takes place of the hollow file created)
4. Create padding to get even 512 blocks - send it to stdout
5. Remove both lzo file and hollow file

This sequence is done in subshell whose output is then piped to
compression algorithms.

This way we do not need to have additional data on filesystem (apart
from the small hollow file) while we can create identical archive as
previously.